use doubletime

+++ Assets/Scripts/Behaviours/PedAI/PathMovementData.cs
@@ -8,14 +8,14 @@ namespace SanAndreasUnity.Behaviours.Peds.AI
         public PathNode? currentNode;
         public PathNode? destinationNode;
         public Vector3 moveDestination;
-        public float timeWhenAttemptedToFindClosestNode = 0f;
+        public double timeWhenAttemptedToFindClosestNode = 0;
 
         public void Cleanup()
         {
             this.currentNode = null;
             this.destinationNode = null;
             this.moveDestination = Vector3.zero;
-            this.timeWhenAttemptedToFindClosestNode = 0f;
+            this.timeWhenAttemptedToFindClosestNode = 0;
         }
     }
 }
